Cambio de entero a floltante en el ajuste de sueldo en la logica del financiero

+++ Corpoagrima/src/main/java/corpoagrima/corpoagrima/logic/DatoEstadoFinanciero.java
@@ -71,7 +71,7 @@ public class DatoEstadoFinanciero {
 
         while (empleadoResult.next()) {
             sueldoBase += empleadoResult.getFloat("Salario_base");
-            ajustoSueldo += Integer.valueOf(empleadoResult.getString("Ajuste_Sueldo"));
+            ajustoSueldo += Double.valueOf(empleadoResult.getString("Ajuste_Sueldo"));
             bonificacion += empleadoResult.getInt("Bonificaciones");
         }
